requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
requests>=2.28.0
pydantic>=2.0.0
pytz>=2022.0
langchain-deepseek>=0.0.2
gunicorn>=20.1.0
sentry-sdk[flask]>=1.39.0
python-dateutil>=2.8.0
stripe>=7.0.0